Extract FacetTagsPresenter out of ResultSetPresenter
This moves the facet tags presentation logic into its own class.
