+++ app/controllers/finders_controller.rb
@@ -21,25 +21,17 @@ class FindersController < ApplicationController
       end
       format.json do
         if content_item.is_search? || content_item.is_finder?
-          render json: {
-            total: results.displayed_total,
-            facet_tags: render_component("facet_tags", results.facet_tags_content),
-            search_results: render_component("finders/search_results", results.search_results_content),
-            sort_options_markup: render_component("finders/sort_options", sort_presenter.to_hash),
-            next_and_prev_links: render_component("govuk_publishing_components/components/previous_and_next_navigation", results.next_and_prev_links),
-          }
+          render json: json_response
         else
           render json: {}, status: :not_found
         end
       end
       format.atom do
         if content_item.is_redirect?
-          @redirect = content_item.as_hash.dig('redirects', 0, 'destination')
-          @finder_slug = finder_slug
-          render 'finders/show-redirect'
+          redirect_to_destination
         elsif finder.atom_feed_enabled?
           expires_in(ATOM_FEED_MAX_AGE, public: true)
-          @feed = AtomPresenter.new(finder, results)
+          @feed = AtomPresenter.new(finder, results, facet_tags)
         else
           render plain: 'Not found', status: :not_found
         end
@@ -50,6 +42,24 @@ class FindersController < ApplicationController
   end
 
 private
+
+  helper_method :finder, :facet_tags
+
+  def redirect_to_destination
+    @redirect = content_item.as_hash.dig('redirects', 0, 'destination')
+    @finder_slug = finder_slug
+    render 'finders/show-redirect'
+  end
+
+  def json_response
+    {
+      total: results.displayed_total,
+      facet_tags: render_component("facet_tags", facet_tags.present),
+      search_results: render_component("finders/search_results", results.search_results_content),
+      sort_options_markup: render_component("finders/sort_options", sort_presenter.to_hash),
+      next_and_prev_links: render_component("govuk_publishing_components/components/previous_and_next_navigation", results.next_and_prev_links),
+    }
+  end
 
   def render_component(partial, locals)
     (render_to_string(formats: %w[html], partial: partial, locals: locals) || "").squish
@@ -78,7 +88,6 @@ private
       filter_params,
     )
   end
-  helper_method :finder
 
   def finder_api
     @finder_api ||= finder_api_class.new(content_item.as_hash, filter_params)
@@ -118,6 +127,10 @@ private
 
   def parent
     params.fetch(:parent, '')
+  end
+
+  def facet_tags
+    @facet_tags ||= FacetTagsPresenter.new(finder, sort_presenter)
   end
 
   def grouped_display?
