+++ app/views/finders/show.html.erb
@@ -110,7 +110,7 @@
             <% end %>
           </div>
           <div id="js-facet-tag-wrapper">
-            <%= render "facet_tags", @results.facet_tags_content %>
+            <%= render "facet_tags", facet_tags.present %>
           </div>
         </div>
       </div>
