+++ spec/presenters/advanced_search_result_set_presenter_spec.rb
@@ -51,61 +51,4 @@ RSpec.describe AdvancedSearchResultSetPresenter do
     allow(Services.rummager).to receive(:batch_search)
       .and_return("results" => [search_results])
   end
-
-  describe "#facet_tags_content" do
-    context "applied filters, filtered with dates" do
-      let(:filter_params) {
-        {
-          "topic" => "/education",
-          "group" => "news_and_communications",
-          "public_timestamp" => { "from" => "02-2017" },
-        }
-      }
-
-      it "contains all subgroups and date filters" do
-        expected = [[{
-          data_facet: "public_timestamp",
-          data_name: "public_timestamp[from]",
-          data_track_label: " 1 February 2017",
-          data_value: "02-2017",
-          preposition: "Published After",
-          text: " 1 February 2017"
-        }]]
-        expect(instance.facet_tags_content[:applied_filters]).to eq(expected)
-      end
-    end
-
-    context "applied filters, filtered with dates and subgroups" do
-      let(:filter_params) {
-        {
-          "topic" => "/education",
-          "group" => "news_and_communications",
-          "subgroup" => "news",
-          "public_timestamp" => { "from" => "02-2017" },
-        }
-      }
-
-      it "contains selected subgroups and date filters" do
-        expected = [
-          [{
-            data_facet: "subgroup",
-            data_name: nil,
-            data_track_label: "News",
-            data_value: "news",
-            preposition: "In",
-            text: "News"
-          }],
-          [{
-            data_facet: "public_timestamp",
-            data_name: "public_timestamp[from]",
-            data_track_label: " 1 February 2017",
-            data_value: "02-2017",
-            preposition: "Published After",
-            text: " 1 February 2017"
-          }]
-        ]
-        expect(instance.facet_tags_content[:applied_filters]).to eq(expected)
-      end
-    end
-  end
 end
