+++ spec/presenters/result_set_presenter_spec.rb
@@ -1,6 +1,9 @@
 require 'spec_helper'
+require_relative './helpers/facets_helper'
 
 RSpec.describe ResultSetPresenter do
+  include FacetsHelper
+
   subject(:presenter) { ResultSetPresenter.new(finder, filter_params, sort_presenter, metadata_presenter_class) }
   let(:metadata_presenter_class) do
     MetadataPresenter
@@ -46,102 +49,10 @@ RSpec.describe ResultSetPresenter do
 
   let(:filter_params) { { keywords: 'test' } }
 
-  let(:a_facet) do
-    double(
-      OptionSelectFacet,
-      key: 'key_1',
-      selected_values: [
-      {
-        'value' => 'ca98-and-civil-cartels',
-        'label' => 'CA98 and civil cartels'
-      },
-      {
-        'value' => 'mergers',
-        'label' => 'Mergers'
-      },
-    ],
-      allowed_values: [
-      {
-        'value' => 'ca98-and-civil-cartels',
-        'label' => 'CA98 and civil cartels'
-      },
-      {
-        'value' => 'mergers',
-        'label' => 'Mergers'
-      },
-    ],
-      sentence_fragment: {
-      'key' => 'key_1',
-      'type' => 'text',
-      'preposition' => 'Of Type',
-      'values' => [
-        {
-          'label' => 'CA98 and civil cartels',
-        },
-        {
-          'label' => 'Mergers',
-        },
-      ],
-      'word_connectors' => { words_connector: 'or' }
-    },
-      has_filters?: true,
-      labels: %W(ca98-and-civil-cartels mergers),
-      value: %W(ca98-and-civil-cartels mergers),
-      hide_facet_tag?: false
-    )
-  end
-
   let(:a_facet_collection) do
     double(
       FacetCollection,
       filters: [a_facet, another_facet, a_date_facet]
-    )
-  end
-
-  let(:another_facet) do
-    double(
-      OptionSelectFacet,
-      key: 'key_2',
-      preposition: 'About',
-      selected_values: [
-      {
-        'value' => 'farming',
-        'label' => 'Farming'
-      },
-      {
-        'value' => 'chemicals',
-        'label' => 'Chemicals'
-      },
-    ],
-      sentence_fragment: {
-      'key' => 'key_2',
-      'type' => 'text',
-      'preposition' => 'About',
-      'values' => [
-        {
-          'label' => 'Farming',
-        },
-        {
-          'label' => 'Chemicals',
-        },
-      ],
-      'word_connectors' => { words_connector: 'or' }
-    },
-      has_filters?: true,
-      value: %w[farming chemicals],
-      'word_connectors' => { words_connector: 'or' },
-      hide_facet_tag?: false
-    )
-  end
-
-  let(:a_date_facet) do
-    double(
-      OptionSelectFacet,
-      'key' => 'closed_date',
-      sentence_fragment: nil,
-      has_filters?: false,
-      'word_connectors' => { words_connector: 'or' },
-      hide_facet_tag?: false
     )
   end
 
@@ -200,38 +111,6 @@ RSpec.describe ResultSetPresenter do
     )
   end
 
-  let(:a_facet_without_facet_tags) do
-    double(
-      RadioFacet,
-      key: 'key_3',
-      preposition: 'that are',
-      allowed_values: [
-      {
-        'value' => 'statistics_published',
-        'label' => 'Statistics (published)',
-        'default' => true
-      },
-      {
-        'value' => 'statistics_upcoming',
-        'label' => 'Statistics (upcoming)'
-      },
-      {
-        'value' => 'research',
-        'label' => 'Research'
-      },
-    ],
-      has_filters?: true,
-      hide_facet_tag?: true,
-      value: "something",
-      sort: [
-      {
-        'value' => 'most-viewed',
-        'name' => 'Most viewed'
-      }
-    ]
-    )
-  end
-
   before(:each) do
     allow(finder).to receive(:eu_exit_finder?).and_return(false)
   end
@@ -239,59 +118,6 @@ RSpec.describe ResultSetPresenter do
   describe "#displayed_total" do
     it 'combines total with document noun' do
       expect(presenter.displayed_total).to eql("#{total} cases")
-    end
-  end
-
-  describe '#selected_filter_descriptions' do
-    before(:each) do
-      allow(presenter).to receive(:link_without_facet_value)
-      allow(finder).to receive(:filters).and_return([a_facet, another_facet, a_date_facet])
-    end
-
-    it 'includes prepositions for each facet' do
-      applied_filters = presenter.selected_filter_descriptions.flat_map { |filter| filter }
-      prepositions = applied_filters.flat_map { |filter| filter[:preposition] }.reject { |preposition| preposition == "or" }
-
-      finder.filters.reject { |filter| filter.sentence_fragment.nil? }.each do |fragment|
-        expect(prepositions).to include(fragment.sentence_fragment['preposition'])
-      end
-    end
-
-    context 'when keywords have been searched for' do
-      let(:keywords) { "my search term" }
-
-      it 'includes the keywords' do
-        applied_filters = presenter.selected_filter_descriptions.flat_map { |filter| filter }
-        text_values = applied_filters.flat_map { |filter| filter[:text] }
-
-        expect(text_values).to include("my", "search", "term")
-      end
-    end
-
-    context 'when XSS attack keywords have been searched for' do
-      let(:keywords) { '<script>alert("hello")</script>' }
-
-      it 'escapes keywords appropriately' do
-        applied_filters = presenter.selected_filter_descriptions.flat_map { |filter| filter }
-        text_values = applied_filters.flat_map { |filter| filter[:text] }
-
-        expect(["script", "alert", "&quot;hello&quot;"].any? { |word| text_values.join(" ").include?(word) })
-      end
-    end
-  end
-
-  describe '#facet_values_sentence' do
-    before(:each) do
-      allow(finder).to receive(:filters).and_return([a_facet, another_facet, a_date_facet])
-    end
-
-    let(:applied_filters) { presenter.selected_filter_descriptions.flat_map { |filter| filter } }
-
-    it 'returns an array of hashes that can be used to construct facet tags' do
-      text_values = applied_filters.flat_map { |filter| filter[:text] }
-      finder.facets.filters.flat_map { |filter| filter.sentence_fragment.nil? ? nil : filter.sentence_fragment['values'] }.compact.each do |value|
-        expect(text_values).to include(value['label'])
-      end
     end
   end
